+++ fltk-sys/src/fl.rs
@@ -456,3 +456,12 @@ extern "C" {
         arg5: libc::c_int,
     );
 }
+extern "C" {
+    pub fn Fl_draw_box_active() -> libc::c_int;
+}
+extern "C" {
+    pub fn Fl_gray_ramp(i: libc::c_int) -> libc::c_uint;
+}
+extern "C" {
+    pub fn Fl_color_average(arg1: libc::c_uint, arg2: libc::c_uint, f: f32) -> libc::c_uint;
+}